fixed some errors relating redux in register page and changed redux file structure a bit
