+++ frontend/src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { Provider } from "react-redux";
-import { store } from "../redux/index";
+import { store } from "../redux/store";
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
   return (
