--- frontend/src/redux/action-creators/action_types.ts
@@ -1,6 +1,0 @@
-export enum ActionType {
-  LOGIN_POST = "LOGIN_POST",
-  LOGIN_POST_ERROR = "LOGIN_POST_ERROR",
-  REGISTER_POST = "REGISTER_POST",
-  REGISTER_POST_ERROR = "REGISTER_POST_ERROR",
-}
