--- frontend/src/redux/index.ts
@@ -1,3 +1,0 @@
-export * as actionCreators from "./action-creators";
-export * from "./store";
-export * from "./reducer/index";
